runtime/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agency-runtime"
version = "0.1.0"
description = "Orchestration runtime that loads The Agency persona files as runnable skills."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
  "anthropic>=0.39.0",
  "fastapi>=0.110.0",
  "uvicorn>=0.27.0",
  "pydantic>=2.6.0",
  "click>=8.1.0",
  "PyYAML>=6.0",
  "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23"]
docs = [
  "pypdf>=4.0",
  "python-docx>=1.1",
  "openpyxl>=3.1",
]
computer = [
  "pyautogui>=0.9.54",
  "pillow>=10.0",
]

[project.scripts]
agency = "agency.cli:main"

[tool.setuptools.packages.find]
include = ["agency*"]

[tool.setuptools.package-data]
agency = ["static/*.html", "static/*.css", "static/*.js"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
